requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pandas==2.0.3
numpy==1.24.3
openai==1.3.5
python-dotenv==1.0.0
openpyxl==3.1.2
scikit-learn==1.3.0
tqdm==4.66.1
langdetect==1.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2
transformers==4.33.2
torch==2.0.1
# First, uninstall the problematic packages
python -m pip uninstall sentence-transformers huggingface_hub transformers -y

# Install compatible versions
python -m pip install huggingface_hub==0.16.4
python -m pip install transformers==4.33.2
python -m pip install sentence-transformers==2.2.2

# Test again
python quick_test.py